# p2p/gen/openapi/bidderapi/v1/bidderapi.swagger.yaml
swagger: "2.0"
info:
  title: Bidder API
  version: 1.0.0-alpha
  license:
    name: Business Source License 1.1
    url: https://github.com/primevprotocol/mev-commit/blob/main/LICENSE
consumes:
  - application/json
produces:
  - application/json
paths:
  /v1/bidder/bid:
    post:
      summary: SendBid
      description: Send a bid to the bidder mev-commit node.
      operationId: Bidder_SendBid
      responses:
        "200":
          description: A successful response.(streaming responses)
          schema:
            type: object
            properties:
              result:
                $ref: '#/definitions/v1Commitment'
              error:
                $ref: '#/definitions/googlerpcStatus'
            title: Stream result of v1Commitment
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/googlerpcStatus'
      parameters:
        - name: body
          description: Unsigned bid message from bidders to the bidder mev-commit node.
          in: body
          required: true
          schema:
            $ref: '#/definitions/bidderapiv1Bid'
  /v1/bidder/get_allowance:
    get:
      summary: GetAllowance
      description: GetAllowance is called by the bidder to get its allowance in the bidder registry.
      operationId: Bidder_GetAllowance
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/v1PrepayResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/googlerpcStatus'
  /v1/bidder/get_min_allowance:
    get:
      summary: GetMinAllowance
      description: GetMinAllowance is called by the bidder to get the minimum allowance required in the bidder registry to make bids.
      operationId: Bidder_GetMinAllowance
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/v1PrepayResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/googlerpcStatus'
  /v1/bidder/prepay/{amount}:
    post:
      summary: PrepayAllowance
      description: PrepayAllowance is called by the bidder node to add prepaid allowance in the bidder registry.
      operationId: Bidder_PrepayAllowance
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/v1PrepayResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/googlerpcStatus'
      parameters:
        - name: amount
          description: Amount of ETH to be prepaid in wei.
          in: path
          required: true
          type: string
definitions:
  bidderapiv1Bid:
    type: object
    example:
      amount: "1000000000000000000"
      blockNumber: 123456
      txHashes:
        - fe4cb47db3630551beedfbd02a71ecc69fd59758e2ba699606e2d5c74284ffa7
        - 71c1348f2d7ff7e814f9c3617983703435ea7446de420aeac488bf1de35737e8
    properties:
      txHashes:
        type: array
        items:
          type: string
          pattern: '[a-fA-F0-9]{64}'
        description: Hex string encoding of the hashes of the transactions that the bidder wants to include in the block.
      amount:
        type: string
        description: Amount of ETH that the bidder is willing to pay to the provider for including the transaction in the block.
        pattern: '[0-9]+'
      blockNumber:
        type: string
        format: int64
        description: Max block number that the bidder wants to include the transaction in.
      decayStartTimestamp:
        type: string
        format: int64
        description: Timestamp at which the bid starts decaying.
      decayEndTimestamp:
        type: string
        format: int64
        description: Timestamp at which the bid ends decaying.
    description: Unsigned bid message from bidders to the bidder mev-commit node.
    title: Bid message
    required:
      - txHashes
      - amount
      - blockNumber
  googlerpcStatus:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      details:
        type: array
        items:
          type: object
          $ref: '#/definitions/protobufAny'
  protobufAny:
    type: object
    properties:
      '@type':
        type: string
    additionalProperties: {}
  v1Commitment:
    type: object
    properties:
      txHashes:
        type: array
        items:
          type: string
          pattern: '[a-fA-F0-9]{64}'
        description: Hex string encoding of the hash of the transaction that the bidder wants to include in the block.
      bidAmount:
        type: string
        description: Amount of ETH that the bidder has agreed to pay to the provider for including the transaction in the block.
      blockNumber:
        type: string
        format: int64
        description: Max block number that the bidder wants to include the transaction in.
      receivedBidDigest:
        type: string
        description: Hex string encoding of digest of the bid message signed by the bidder.
      receivedBidSignature:
        type: string
        description: Hex string encoding of signature of the bidder that sent this bid.
      commitmentDigest:
        type: string
        description: Hex string encoding of digest of the commitment.
      commitmentSignature:
        type: string
        description: Hex string encoding of signature of the commitment signed by the provider confirming this transaction.
      providerAddress:
        type: string
        description: Hex string encoding of the address of the provider that signed the commitment signature.
      decayStartTimestamp:
        type: string
        format: int64
        description: Timestamp at which the bid starts decaying.
      decayEndTimestamp:
        type: string
        format: int64
        description: Timestamp at which the bid ends decaying.
  v1PrepayResponse:
    type: object
    example:
      amount: "1000000000000000000"
    properties:
      amount:
        type: string
    description: Get prepaid allowance for bidder in the bidder registry.
    title: Prepay response